Skip infer test in windows
